+++ run_linters.sh
@@ -26,10 +26,13 @@ check_log "cppcheck  find_max_incr_sequence_lib/ main.c --enable=all --inconclus
         -I  find_max_incr_sequence_lib/ --suppress=missingIncludeSystem" "\(information\)"
 
 print_header "RUN clang-tidy"
-check_log "clang-tidy find_max_incr_sequence_lib/*.c find_max_incr_sequence_lib/*.h \
-        -warnings-as-errors=* -extra-arg=-std=c99 -- -Ifind_max_incr_sequence_lib/" "Error (?:reading|while processing)"
+check_log "clang-tidy find_max_incr_sequence_lib/*.c find_max_incr_sequence_lib/array_lib/*.c \
+        find_max_incr_sequence_lib/*.h find_max_incr_sequence_lib/array_lib/*.h \
+        -warnings-as-errors=* -extra-arg=-std=gnu99 -- -Ifind_max_incr_sequence_lib/ \
+        -Ifind_max_incr_sequence_lib/array_lib/" "Error (?:reading|while processing)"
 
 print_header "RUN cpplint"
-check_log "cpplint --extensions=c find_max_incr_sequence_lib/*" "Can't open for reading"
+check_log "cpplint --extensions=c find_max_incr_sequence_lib/*.c find_max_incr_sequence_lib/*.h \
+        find_max_incr_sequence_lib/array_lib/*" "Can't open for reading"
 
 print_header "SUCCESS"
